Move loader stuff around
