+++ package.json
@@ -103,6 +103,7 @@
     "jest-environment-jsdom-fourteen": "^1.0.1",
     "jest-fetch-mock": "^3.0.3",
     "jest-image-snapshot": "^3.1.0",
+    "jest-localstorage-mock": "^2.4.3",
     "jest-pnp-resolver": "^1.2.0",
     "lerna": "^3.4.1",
     "mobx": "^5.0.0",
