+++ products/jbrowse-web/src/Loader.tsx
@@ -208,6 +208,10 @@ export function Loader() {
         const sessionStorageSession = JSON.parse(
           sessionStorage.getItem('current') || '{}',
         )
+        console.log(
+          localStorage.getItem('test'),
+          sessionStorage.getItem('test'),
+        )
         const foundLocalSession =
           localStorage.getItem(sessionQueryParam) ||
           (sessionStorageSession.id === sessionQueryParam
